Renamed registerAccountModal to contactInfoModal
