+++ src/common/components/contactInfoModal/contactInfoModal.tpl.html
