+++ src/common/services/session/sessionModal.controller.js
@@ -5,7 +5,7 @@ import signUpModal from 'common/components/signUpModal/signUpModal.component';
 import resetPasswordModal from 'common/components/resetPasswordModal/resetPasswordModal.component';
 import forgotPasswordModal from 'common/components/forgotPasswordModal/forgotPasswordModal.component';
 import userMatchModal from 'common/components/userMatchModal/userMatchModal.component';
-import registerAccountModal from 'common/components/registerAccountModal/registerAccountModal.component';
+import contactInfoModal from 'common/components/contactInfoModal/contactInfoModal.component';
 import accountBenefitsModal from 'common/components/accountBenefitsModal/accountBenefitsModal.component';
 
 let controllerName = 'sessionModalController';
@@ -52,7 +52,7 @@ export default angular
     resetPasswordModal.name,
     forgotPasswordModal.name,
     userMatchModal.name,
-    registerAccountModal.name,
+    contactInfoModal.name,
     accountBenefitsModal.name
   ] )
   .controller( controllerName, SessionModalController );
